pyproject.toml
[project]
name = "video2slider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.33.0",
    "python-dotenv>=1.2.1",
]
